Split tests in files
